requirements.txt
# Flask REST API (src/ layout)
Flask>=3.0.0,<4
flask-cors>=4.0.0
SQLAlchemy>=2.0.0,<3
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.0.0
redis>=5.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
requests>=2.31.0
bcrypt>=4.1.0
